src/main/webapp/app/entities/audit-log/route/audit-log-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IAuditLog, AuditLog } from '../audit-log.model';
import { AuditLogService } from '../service/audit-log.service';

import { AuditLogRoutingResolveService } from './audit-log-routing-resolve.service';

describe('Service Tests', () => {
  describe('AuditLog routing resolve service', () => {
    let mockRouter: Router;
    let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
    let routingResolveService: AuditLogRoutingResolveService;
    let service: AuditLogService;
    let resultAuditLog: IAuditLog | undefined;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        providers: [Router, ActivatedRouteSnapshot],
      });
      mockRouter = TestBed.inject(Router);
      mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
      routingResolveService = TestBed.inject(AuditLogRoutingResolveService);
      service = TestBed.inject(AuditLogService);
      resultAuditLog = undefined;
    });

    describe('resolve', () => {
      it('should return IAuditLog returned by find', () => {
        // GIVEN
        service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
        mockActivatedRouteSnapshot.params = { id: 123 };

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultAuditLog = result;
        });

        // THEN
        expect(service.find).toBeCalledWith(123);
        expect(resultAuditLog).toEqual({ id: 123 });
      });

      it('should return new IAuditLog if id is not provided', () => {
        // GIVEN
        service.find = jest.fn();
        mockActivatedRouteSnapshot.params = {};

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultAuditLog = result;
        });

        // THEN
        expect(service.find).not.toBeCalled();
        expect(resultAuditLog).toEqual(new AuditLog());
      });

      it('should route to 404 page if data not found in server', () => {
        // GIVEN
        spyOn(service, 'find').and.returnValue(of(new HttpResponse({ body: null })));
        mockActivatedRouteSnapshot.params = { id: 123 };

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultAuditLog = result;
        });

        // THEN
        expect(service.find).toBeCalledWith(123);
        expect(resultAuditLog).toEqual(undefined);
        expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
      });
    });
  });
});
